Move all components to components folder
