+++ src/index.svelte
@@ -1,7 +1,7 @@
 <script>
-  import Presentation from './Presentation.svelte';
-  import Header from './Header.svelte';
-  import Footer from './Footer.svelte';
+  import Header from './components/Header.svelte';
+  import Presentation from './components/Presentation.svelte';
+  import Footer from './components/Footer.svelte';
 </script>
 
 <main class="h-screen flex flex-col justify-between bg-gray-800 text-gray-600">
