+++ src/main.js
@@ -1,6 +1,6 @@
 import { init, getLocaleFromNavigator } from 'svelte-i18n';
 import App from './index.svelte';
-import './Tailwind.svelte';
+import './components/Tailwind.svelte';
 import './i18n';
 
 const fallbackLocale = 'en-US';
